requirements.txt
numpy==1.26.4
Pillow
black
pre-commit
pyright
pylint
